requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
django-cors-headers>=4.3
mysqlclient>=2.2         # for MySQL in production (comment out if using SQLite)
gunicorn>=21.2           # for production deployment
python-decouple>=3.8     # for environment variable management
